Checks the overall status of the R session

Individual session check functions that each inspect one way in which an R session could be considered not to be "clean". Session checkers can produce errors, warnings, or messages if requested.

Usage

check_session(
  action = "warn",
  check_globalenv = TRUE,
  check_packages = TRUE,
  check_namespaces = TRUE,
  check_attachments = TRUE,
  ...,
  settings = getOption("sessioncheck.settings")
)

Arguments

action

Behaviour to take if the status is not clean. Possible values are "error", "warn", "message", and "none". The default is action = "warn".

check_globalenv

Should the check_globalenv() check be run? (default is TRUE)

check_packages

Should the check_packages() check be run? (default is TRUE)

check_namespaces

Should the check_namespaces() check be run? (default is TRUE)

check_attachments

Should the check_attachments() check be run? (default is TRUE)

...

Reserved to allow additional checks to be added in future (ignored)

settings

A list that specifies the rules applied for individual checks

Value

Invisibly returns a status object, a list of a named logical vectors. Each vector has names that refer to detected entities for each specific check. Values are TRUE if the entity is ignored, FALSE if it triggers an action.

Details

check_session() allows the user to apply multiple session checks in a single function

Examples

check_session(action = "message")
#> Found attached packages: sessioncheck
#> Found loaded namespaces: jsonlite, xml2, jquerylib, textshaping, and 39 more